+++ monopyly/auth/__init__.py
@@ -7,4 +7,4 @@ from flask import Blueprint
 # Define the blueprint
 auth = Blueprint('auth', __name__, url_prefix='/auth')
 
-import monopyly.auth.routes
+from . import routes