+++ monopyly/core/__init__.py
@@ -6,6 +6,6 @@ from flask import Blueprint
 # Define the blueprint
 core = Blueprint('core', __name__)
 
-import monopyly.core.routes
-import monopyly.core.filters
-import monopyly.core.context_processors
+from . import routes
+from . import filters
+from . import context_processors