+++ monopyly/credit/__init__.py
@@ -7,4 +7,4 @@ from flask import Blueprint
 # Define the blueprint
 credit = Blueprint('credit', __name__, url_prefix='/credit')
 
-import monopyly.credit.routes
+from . import routes