+++ monopyly/credit/transactions.py
@@ -4,11 +4,11 @@ Tools for interacting with the credit transactions in the database.
 import datetime
 from sqlite3 import IntegrityError
 
+from ..db import DATABASE_FIELDS
 from ..utils import (
-    DatabaseHandler, fill_places, filter_items, filter_dates, check_sort_order
+    DatabaseHandler, fill_places, filter_items, filter_dates, check_sort_order,
+    select_fields
 )
-from .constants import TRANSACTION_FIELDS, TAG_FIELDS
-from .tools import select_fields
 
 
 class TransactionHandler(DatabaseHandler):
@@ -28,7 +28,7 @@ class TransactionHandler(DatabaseHandler):
 
     Attributes
     ––––––––––
-    table_name : str
+    table : str
         The name of the database table that this handler manages.
     db : sqlite3.Connection
         A connection to the database for interfacing.
@@ -37,14 +37,14 @@ class TransactionHandler(DatabaseHandler):
     user_id : int
         The ID of the user who is the subject of database access.
     """
-    table_name = 'credit_transactions'
-    table_fields = TRANSACTION_FIELDS
+    _table = 'credit_transactions'
+    _table_view = 'credit_transactions_view'
 
     def __init__(self, db=None, user_id=None, check_user=True):
         super().__init__(db=db, user_id=user_id, check_user=check_user)
 
     def get_entries(self, card_ids=None, statement_ids=None, active=False,
-                    sort_order='DESC', fields=TRANSACTION_FIELDS):
+                    sort_order='DESC', fields=DATABASE_FIELDS[_table_view]):
         """
         Get credit card transactions from the database.
 
@@ -87,7 +87,7 @@ class TransactionHandler(DatabaseHandler):
         statement_filter = filter_items(statement_ids, 'statement_id', 'AND')
         active_filter = "AND active = 1" if active else ""
         query = (f"SELECT {select_fields(fields, 't.id')} "
-                  "  FROM credit_transactions AS t "
+                  "  FROM credit_transactions_view AS t "
                   "       INNER JOIN credit_statements AS s "
                   "          ON s.id = t.statement_id "
                   "       INNER JOIN credit_cards AS c "
@@ -96,6 +96,7 @@ class TransactionHandler(DatabaseHandler):
                   "          ON a.id = c.account_id "
                   " WHERE user_id = ? "
                  f"       {card_filter} {statement_filter} {active_filter} "
+                  " GROUP BY t.id "
                  f" ORDER BY transaction_date {sort_order}")
         placeholders = (self.user_id, *fill_places(card_ids),
                         *fill_places(statement_ids))
@@ -104,7 +105,7 @@ class TransactionHandler(DatabaseHandler):
 
     def get_entry(self, transaction_id, fields=None):
         """
-        Get a transaction from the database given its transaction ID.
+        Get a transaction from the database given its ID.
 
         Accesses a set of fields for a given transaction. By default,
         all fields for a transaction, the corresponding statement,
@@ -124,7 +125,7 @@ class TransactionHandler(DatabaseHandler):
             The transaction information from the database.
         """
         query = (f"SELECT {select_fields(fields, 't.id')} "
-                  "  FROM credit_transactions AS t "
+                  "  FROM credit_transactions_view AS t "
                   "       INNER JOIN credit_statements AS s "
                   "          ON s.id = t.statement_id "
                   "       INNER JOIN credit_cards AS c "
@@ -132,21 +133,215 @@ class TransactionHandler(DatabaseHandler):
                   "       INNER JOIN credit_accounts AS a "
                   "          ON a.id = c.account_id "
                   " WHERE t.id = ? AND user_id = ?")
+        placeholders = (transaction_id, self.user_id)
         abort_msg = (f'Transaction ID {transaction_id} does not exist for the '
                       'user.')
-        transaction = self._query_entry(transaction_id, query, abort_msg)
+        transaction = self._query_entry(query, placeholders, abort_msg)
         return transaction
 
-    def add_transaction(self, statement_id, transaction_date, vendor, amount,
-                        notes):
-        """Add a transaction to the database."""
-        transaction_data = {'statement_id': statement_id,
-                            'transaction_date': transaction_date,
-                            'vendor': vendor,
-                            'amount': amount,
-                            'notes': notes}
-        transaction = self.add_entry(transaction_data)
-        return transaction
+    def add_entry(self, mapping):
+        """
+        Add a transaction to the database.
+
+        Uses a mapping produced by a `TransactionForm` to add a new
+        transaction into the database. The mapping includes information
+        for the transaction, along with information for all
+        subtransactions (including tags associated with each
+        subtransaction).
+
+        Parameters
+        ––––––––––
+        mapping : dict
+            A mapping between database fields and the value to be
+            entered into that field for the transaction. The mapping
+            also contains subtransaction information (including tags).
+
+        Returns
+        –––––––
+        transaction : sqlite3.Row
+            The saved transaction.
+        subtransactions : list of sqlite3.Row
+            A list of subtransactions belonging to the saved transaction.
+        """
+        subtransaction_db = SubtransactionHandler()
+        # Override the default method to account for subtransactions
+        subtransactions_data = mapping.pop('subtransactions')
+        transaction = super().add_entry(mapping)
+        subtransactions = self._add_subtransactions(transaction['id'],
+                                                    subtransactions_data)
+        # Refresh the transaction information
+        transaction = self.get_entry(transaction['id'])
+        return transaction, subtransactions
+
+    def update_entry(self, entry_id, mapping):
+        """Update a transaction (and its subtransactions) in the database."""
+        subtransaction_db = SubtransactionHandler()
+        # Override the default method to account for subtransactions
+        subtransactions_data = mapping.pop('subtransactions')
+        transaction = super().update_entry(entry_id, mapping)
+        # Replace subtransactions when updating
+        subtransactions = subtransaction_db.get_entries((transaction['id'],))
+        subtransaction_db.delete_entries(
+            [subtransaction['id'] for subtransaction in subtransactions]
+        )
+        subtransactions = self._add_subtransactions(transaction['id'],
+                                                    subtransactions_data)
+        # Refresh the transaction information
+        transaction = self.get_entry(transaction['id'])
+        return transaction, subtransactions
+
+    def _add_subtransactions(self, transaction_id, subtransactions_data):
+        """Add subtransactions to the database for the data given."""
+        subtransaction_db = SubtransactionHandler()
+        # Assemble mappings to add subtransactions
+        subtransactions = []
+        for subtransaction_data in subtransactions_data:
+            # Complete the mapping and add the subtransaction to the database
+            sub_mapping = {'transaction_id': transaction_id,
+                           **subtransaction_data}
+            subtransaction = subtransaction_db.add_entry(sub_mapping)
+            subtransactions.append(subtransaction)
+        return subtransactions
+
+
+class SubtransactionHandler(DatabaseHandler):
+    """
+    A database handler for accessing credit subtransactions.
+
+    Parameters
+    ––––––––––
+    db : sqlite3.Connection
+        A connection to the database for interfacing.
+    user_id : int
+        The ID of the user who is the subject of database access. If not
+        given, the handler defaults to using the logged-in user.
+    check_user : bool
+        A flag indicating whether the handler should check that the
+        provided user ID matches the logged-in user.
+
+    Attributes
+    ––––––––––
+    table : str
+        The name of the database table that this handler manages.
+    db : sqlite3.Connection
+        A connection to the database for interfacing.
+    cursor : sqlite.Cursor
+        A cursor for executing database interactions.
+    user_id : int
+        The ID of the user who is the subject of database access.
+    """
+    _table = 'credit_subtransactions'
+
+    def __init__(self, db=None, user_id=None, check_user=True):
+        super().__init__(db=db, user_id=user_id, check_user=check_user)
+
+    def get_entries(self, transaction_ids=None, fields=None):
+        """
+        Get all subtransactions for a credit transaction.
+
+        Accesses a set of fields for subtransactions belonging to a set
+        of transactions. By default, all fields for the subtransactions
+        are returned.
+
+        Parameters
+        ––––––––––
+        transaction_ids : tuple of int, optional
+            The IDs of the transactions for which to retrieve
+            subtransactions.
+        fields : tuple of str, optional
+            The fields (in either the transactions, subtransactions,
+            statements, cards, or accounts tables) to be returned. By
+            default, all fields are returned.
+
+        Returns
+        –––––––
+        subtransactions : list of sqlite3.Row
+            A list of credit card subtransactions that are associated
+            with the given transaction.
+        """
+        transaction_filter = filter_items(transaction_ids, 'transaction_id',
+                                          'AND')
+        query = (f"SELECT {select_fields(fields, 's_t.id')} "
+                  "  FROM credit_subtransactions AS s_t "
+                  "       INNER JOIN credit_transactions AS t "
+                  "          ON t.id = s_t.transaction_id "
+                  "       INNER JOIN credit_statements AS s "
+                  "          ON s.id = t.statement_id "
+                  "       INNER JOIN credit_cards AS c "
+                  "          ON c.id = s.card_id "
+                  "       INNER JOIN credit_accounts AS a "
+                  "          ON a.id = c.account_id "
+                 f" WHERE user_id = ? {transaction_filter}")
+        placeholders = (self.user_id, *fill_places(transaction_ids))
+        subtransactions = self._query_entries(query, placeholders)
+        return subtransactions
+
+    def get_entry(self, subtransaction_id, fields=None):
+        """
+        Get a subtransaction from the database given its ID.
+
+        Accesses a set of fields for a given subtransaction. By default,
+        all fields for a subtransaction, the corresponding statement,
+        issuing credit card and account are returned.
+
+        Parameters
+        ––––––––––
+        subtransaction_id : int
+            The ID of the subtransaction to be found.
+        fields : tuple of str, optional
+            The fields (in either the subtransactions, transactions,
+            statements, cards, or accounts tables) to be returned.
+
+        Returns
+        –––––––
+        subtransaction : sqlite3.Row
+            The subtransaction information from the database.
+        """
+        query = (f"SELECT {select_fields(fields, 't.id')} "
+                  "  FROM credit_subtransactions AS s_t "
+                  "       INNER JOIN credit_transactions AS t "
+                  "          ON t.id = s_t.transaction_id "
+                  "       INNER JOIN credit_statements AS s "
+                  "          ON s.id = t.statement_id "
+                  "       INNER JOIN credit_cards AS c "
+                  "          ON c.id = s.card_id "
+                  "       INNER JOIN credit_accounts AS a "
+                  "          ON a.id = c.account_id "
+                  " WHERE s_t.id = ? AND user_id = ?")
+        placeholders = (subtransaction_id, self.user_id,)
+        abort_msg = (f'Subtransaction ID {subtransaction_id} does not exist '
+                      'for the user.')
+        subtransaction = self._query_entry(query, placeholders, abort_msg)
+        return subtransaction
+
+    def add_entry(self, mapping):
+        """
+        Add a subtransaction to the database.
+
+        Uses a mapping produced by a `TransactionForm` to add a new
+        subtransaction into the database. The mapping includes
+        information for the subtransaction, including the corresponding
+        transaction. The mapping also provides a list of tags that have
+        been assigned to the new subtransaction.
+
+        Parameters
+        ––––––––––
+        mapping : dict
+            A mapping between database fields and the value to be
+            entered into that field for the subtransaction.
+
+        Returns
+        –––––––
+        subtransaction : sqlite3.Row
+            The saved transaction.
+        """
+        tag_db = TagHandler()
+        # Override the default method to account for tags
+        tags = mapping.pop('tags')
+        subtransaction = super().add_entry(mapping)
+        # Link tags to the subtransaction
+        tag_db.update_tag_links(subtransaction['id'], tags)
+        return subtransaction
 
 
 class TagHandler(DatabaseHandler):
@@ -166,7 +361,7 @@ class TagHandler(DatabaseHandler):
 
     Attributes
     ––––––––––
-    table_name : str
+    table : str
         The name of the database table that this handler manages.
     db : sqlite3.Connection
         A connection to the database for interfacing.
@@ -175,14 +370,14 @@ class TagHandler(DatabaseHandler):
     user_id : int
         The ID of the user who is the subject of database access.
     """
-    table_name = 'credit_tags'
-    table_fields = TAG_FIELDS
+    table = 'credit_tags'
 
     def __init__(self, db=None, user_id=None, check_user=True):
         super().__init__(db=db, user_id=user_id, check_user=check_user)
 
     def get_entries(self, tag_names=None, transaction_ids=None,
-                    fields=TAG_FIELDS):
+                    subtransaction_ids=None, fields=DATABASE_FIELDS[table],
+                    ancestors=True):
         """
         Get credit card transaction tags from the database.
 
@@ -198,10 +393,17 @@ class TagHandler(DatabaseHandler):
         transaction_ids : tuple of int, optional
             A sequence of transaction IDs for which tags will be
             selected (if `None`, all transaction tags will be selected).
+        subtransaction_ids : tuple of int, optional
+            A sequence of subtransaction IDs for which tags will be
+            selected (if `None`, all subtransaction tags will be selected).
         fields : tuple of str, optional
             A sequence of fields to select from the database (if `None`,
             all fields will be selected). A field can be any column from
             the 'credit_tags' table.
+        ancestors : bool, optional
+            A flag indicating whether the query should include tags
+            that are ancestors of other tags in the list of returned
+            tags. The default is `True` (ancestor tags are returned).
 
         Returns
         –––––––
@@ -212,19 +414,32 @@ class TagHandler(DatabaseHandler):
         transaction_filter = filter_items(transaction_ids,
                                           'transaction_id',
                                           'AND')
-        query = (f"SELECT {select_fields(fields, 'DISTINCT t.id')} "
-                  "  FROM credit_tags AS t "
+        subtransaction_filter = filter_items(subtransaction_ids,
+                                             'subtransaction_id',
+                                             'AND')
+        query = (f"SELECT {select_fields(fields, 'DISTINCT tags.id')} "
+                  "  FROM credit_tags AS tags "
                   "       LEFT OUTER JOIN credit_tag_links AS l "
-                  "          ON l.tag_id = t.id "
-                 f" WHERE user_id = ? {name_filter} {transaction_filter}")
+                  "          ON l.tag_id = tags.id "
+                  "       INNER JOIN credit_subtransactions AS s_t "
+                  "          ON s_t.id = l.subtransaction_id "
+                 f" WHERE user_id = ? {name_filter} "
+                 f"       {transaction_filter} {subtransaction_filter}")
         placeholders = (self.user_id, *fill_places(tag_names),
-                        *fill_places(transaction_ids))
+                        *fill_places(transaction_ids),
+                        *fill_places(subtransaction_ids))
         tags = self._query_entries(query, placeholders)
+        # If specified, remove ancestors from the list of tags
+        if not ancestors:
+            for tag in tags:
+                for ancestor in self.get_ancestors(tag['id']):
+                    if ancestor in tags:
+                        tags.remove(ancestor)
         return tags
 
     def get_entry(self, tag_id, fields=None):
         """
-        Get a credit transaction tag from the database given its tag ID.
+        Get a credit transaction tag from the database given its ID.
 
         Accesses a set of fields for a given tag. By default, all fields
         for a tag are returned.
@@ -241,24 +456,25 @@ class TagHandler(DatabaseHandler):
         tag : sqlite3.Row
             The tag information from the database.
         """
-        query = (f"SELECT {select_fields(fields, 't.id')} "
-                  "  FROM credit_tags AS t "
-                  " WHERE t.id = ? AND user_id = ?")
+        query = (f"SELECT {select_fields(fields, 'tags.id')} "
+                  "  FROM credit_tags AS tags "
+                  " WHERE id = ? AND user_id = ?")
+        placeholders = (tag_id, self.user_id)
         abort_msg = (f'Tag ID {tag_id} does not exist for the user.')
-        tag = self._query_entry(tag_id, query, abort_msg)
+        tag = self._query_entry(query, placeholders, abort_msg)
         return tag
 
     def get_subtags(self, tag_id, fields=None):
         """
-        Get subcategories of a credit transaction tag given its tag ID.
+        Get subcategories of a credit transaction tag.
 
         Accesses a set of fields for children of a given tag. By
         default, all fields for a tag are returned.
 
         Parameters
         ––––––––––
-        tag_id : int, None
-            The ID of the tag to be found.
+        tag_id : int
+            The ID of the tag for which to find subtags.
         fields : tuple of str, optional
             The fields (in the tags table) to be returned.
 
@@ -268,11 +484,39 @@ class TagHandler(DatabaseHandler):
             A list of credit card transaction tags that are
             subcategories of the given tag.
         """
-        query = (f"SELECT {select_fields(fields, 't.id')} "
-                  "  FROM credit_tags AS t "
+        query = (f"SELECT {select_fields(fields, 'tags.id')} "
+                  "  FROM credit_tags AS tags "
                   " WHERE parent_id IS ? AND user_id = ?")
-        subtags = self._query_entries(query, (tag_id, self.user_id))
+        placeholders = (tag_id, self.user_id)
+        subtags = self._query_entries(query, placeholders)
         return subtags
+
+    def get_supertag(self, tag_id, fields=None):
+        """
+        Get the supercategory (parent) of a credit transaction tag.
+
+        Accesses a set of fields for the parent of a given tag. By
+        default, all fields for a tag are returned.
+
+        Parameters
+        ––––––––––
+        tag_id : int
+            The ID of the tag for which to find supertags.
+        fields : tuple of str, optional
+            The fields (in the tags table) to be returned.
+
+        Returns
+        –––––––
+        supertag : sqlite3.Row, None
+            The credit card transaction tag that is the parent category
+            of the given tag. Returns `None` if no parent tag is found.
+        """
+        tag = self.get_entry(tag_id)
+        if tag['parent_id']:
+            supertag = self.get_entry(tag['parent_id'])
+        else:
+            supertag = None
+        return supertag
 
     def find_tag(self, tag_name, fields=None):
         """
@@ -294,8 +538,8 @@ class TagHandler(DatabaseHandler):
             The tag entry matching the given criteria. If no matching
             tag is found, returns `None`.
         """
-        query = (f"SELECT {select_fields(fields, 't.id')} "
-                  "  FROM credit_tags AS t "
+        query = (f"SELECT {select_fields(fields, 'tags.id')} "
+                  "  FROM credit_tags AS tags "
                  f" WHERE user_id = ? AND tag_name = ?")
         placeholders = (self.user_id, tag_name)
         tag = self.cursor.execute(query, placeholders).fetchone()
@@ -321,41 +565,63 @@ class TagHandler(DatabaseHandler):
         Returns
         –––––––
         heirarchy : dict
-            The dictionary representing the user's tags. Keys are a
-            tuple of tag IDs and tag names.
+            The dictionary representing the user's tags. Keys are
+            sqlite3.Row objects.
         """
         heirarchy = {}
         for tag in self.get_subtags(parent_id):
-            tag_key = (tag['id'], tag['tag_name'])
-            heirarchy[tag_key] = self.get_heirarchy(tag['id'])
+            heirarchy[tag] = self.get_heirarchy(tag['id'])
         return heirarchy
 
-    def update_tags(self, transaction_id, tag_names):
+    def get_ancestors(self, tag_id):
         """
-        Update the tags for a transaction in the database.
+        Get the ancestor tags of a given tag.
 
-        Given a transaction and a list of tag names, each tag is applied
-        to the transaction. Then, each existing tag that is not in the
-        list of tag names is disassociated with the transaction. Tag
-        names that do not already have database entries are created.
+        Traverses the heirarchy, starting from the given tag and returns
+        a list of all tags that are ancestors of the given tag.
 
         Parameters
         ––––––––––
-        transaction_id : int
-            The ID of a transaction database entry that will be assigned
-            the tags.
-        tag_names : tuple of str
+        tag_id : int
+            The ID of the tag for which to find ancestors.
+
+        Returns
+        –––––––
+        ancestors : list of sqlite3.Row
+            The ancestors of the given tag.
+        """
+        ancestors = []
+        ancestor = self.get_supertag(tag_id, fields=('tag_name',))
+        while ancestor:
+            ancestors.append(ancestor)
+            ancestor = self.get_supertag(ancestor['id'], fields=('tag_name',))
+        return ancestors
+
+    def update_tag_links(self, subtransaction_id, tag_names):
+        """
+        Update the tag links for a transaction in the database.
+
+        Given a subtransaction and a list of tag names, each tag is
+        applied to the subtransaction. Then, each existing tag that is
+        not in the list of tag names is disassociated with the
+        subtransaction. Tag names that do not already have database
+        entries are created.
+
+        Parameters
+        ––––––––––
+        subtransaction_id : int
+            The ID of a subtransaction database entry that will be
+            assigned the tags.
+        tag_names : list of str
             Tag names to be assigned to the given transactions.
         """
-        # Get all of the current tags for the transaction
-        current_tags = self.get_entries(transaction_ids=(transaction_id,))
-        current_tag_names = [tag['tag_name'] for tag in current_tags]
-        # Determine tags to be added and tags to be removed
-        new_tag_names = [name for name in tag_names
-                         if name not in current_tag_names]
-        old_tag_names = [name for name in current_tag_names
-                         if name not in tag_names]
-        for tag_name in new_tag_names:
+        # Remove existing tags
+        for tag in self.get_entries(subtransaction_ids=(subtransaction_id,)):
+            # Unlink the tag from the subtransaction
+            self.unlink(tag['id'], subtransaction_id)
+        # Add the new tags
+        linked_tags = []
+        for tag_name in tag_names:
             # Get the matching tag
             tag = self.find_tag(tag_name, fields=('tag_name',))
             # Create the tag if it does not already exist in the database
@@ -365,35 +631,37 @@ class TagHandler(DatabaseHandler):
                             'tag_name': tag_name}
                 tag = self.add_entry(tag_data)
             # Link the tag to the transaction
-            self.link(tag['id'], transaction_id)
-        for tag_name in old_tag_names:
-            # Get the matching tag
-            tag = self.find_tag(tag_name, fields=())
-            # Unlink the tag from the transaction
-            self.unlink(tag['id'], transaction_id)
+            if tag not in linked_tags:
+                # Link all ancestor tags with this tag
+                for ancestor in self.get_ancestors(tag['id']):
+                    if ancestor not in linked_tags:
+                        self.link(ancestor['id'], subtransaction_id)
+                        linked_tags.append(ancestor)
+                self.link(tag['id'], subtransaction_id)
+                linked_tags.append(tag)
 
-    def link(self, tag_id, transaction_id):
-        """Add a tag to the given transaction."""
-        # Add the transaction-tag link if the two are not already associated
+    def link(self, tag_id, subtransaction_id):
+        """Add a tag to the given subtransaction."""
+        # Add the subtransaction-tag link if the two are not already associated
         try:
             self.cursor.execute(
-                "INSERT INTO credit_tag_links (transaction_id, tag_id) "
+                "INSERT INTO credit_tag_links (subtransaction_id, tag_id) "
                f"     VALUES (?, ?)",
-                (transaction_id, tag_id)
+                (subtransaction_id, tag_id)
             )
             self.db.commit()
         except IntegrityError:
             # The tag link already exists
             pass
 
-    def unlink(self, tag_id, transaction_id):
-        """Remove a tag from the given transaction."""
-        # Delete the transaction-tag link
+    def unlink(self, tag_id, subtransaction_id):
+        """Remove a tag from the given subtransaction."""
+        # Delete the subtransaction-tag link
         self.cursor.execute(
             "DELETE "
             "  FROM credit_tag_links "
-            " WHERE transaction_id = ? AND tag_id = ? ",
-            (transaction_id, tag_id)
+            " WHERE subtransaction_id = ? AND tag_id = ? ",
+            (subtransaction_id, tag_id)
         )
         self.db.commit()
 
@@ -436,7 +704,7 @@ class TagHandler(DatabaseHandler):
         statement_filter = filter_items(statement_ids, 'statement_id', 'AND')
         date_filter = filter_dates(start_date, end_date, 'transaction_date',
                                    'AND')
-        fields = ['SUM(amount) total', 'tag_name']
+        fields = ['SUM(subtotal) total', 'tag_name']
         groups = ['tags.id']
         if group_statements:
             fields.append('t.statement_id')
@@ -445,8 +713,10 @@ class TagHandler(DatabaseHandler):
                   "  FROM credit_tags AS tags "
                   "       INNER JOIN credit_tag_links AS l "
                   "          ON l.tag_id = tags.id "
+                  "       INNER JOIN credit_subtransactions AS s_t "
+                  "          ON s_t.id = l.subtransaction_id "
                   "       INNER JOIN credit_transactions AS t "
-                  "          ON t.id = l.transaction_id "
+                  "          ON t.id = s_t.transaction_id "
                   "       INNER JOIN credit_statements AS s "
                   "          ON s.id = t.statement_id "
                   " WHERE tags.user_id = ? "
@@ -491,12 +761,14 @@ class TagHandler(DatabaseHandler):
                    f" WHERE a.user_id = ? {statement_filter}")
         tag_filter = filter_items(tag_ids, 'tag_id', 'AND')
         statement_filter = filter_items(statement_ids, 'statement_id', 'AND')
-        query = (f"SELECT SUM(amount) / ({subquery}) average_total, tag_name "
+        query = (f"SELECT SUM(subtotal) / ({subquery}) average_total, tag_name "
                  "  FROM credit_tags AS tags "
                  "       INNER JOIN credit_tag_links AS l "
                  "          ON l.tag_id = tags.id "
+                 "       INNER JOIN credit_subtransactions AS s_t "
+                 "          ON s_t.id = l.subtransaction_id "
                  "       INNER JOIN credit_transactions AS t "
-                 "          ON t.id = l.transaction_id "
+                 "          ON t.id = s_t.transaction_id "
                  "       INNER JOIN credit_statements AS s "
                  "          ON s.id = t.statement_id "
                 f" WHERE tags.user_id = ? {tag_filter} {statement_filter} "