+++ monopyly/static/js/expand_transaction.js
@@ -17,19 +17,46 @@ import {
 	// Identify the plus/minus icons
 	const $iconsMoreInfo = $('.transaction .more.button');
 	const $iconsLessInfo = $('.transaction .less.button');
+	// Set timing variables
+	const fadeTime = 200;
+	const slideTime = 250;
+
 
 	$iconsMoreInfo.on('click', function() {
+		// Get the transaction object
 		const $transaction = $(this).closest('.transaction');
-		const endpoint = TRANSACTION_TAGS_ENDPOINT;
+		// Hide the condensed transaction summary
+		const $condensedRow = $transaction.find('.condensed');
+		$condensedRow.fadeTo(fadeTime, 0, function() {
+			$(this).slideUp(slideTime);
+		});
+		// Execute the AJAX request to get transaction/subtransaction information
+		const endpoint = EXPAND_TRANSACTION_ENDPOINT;
 		const rawData = $transaction[0].id;
-		const $tagsContainer = $transaction.find('.tags-container');
-		replaceDisplayContentsAjaxRequest(endpoint, rawData, $tagsContainer);
-		$transaction.addClass('expanded');
+		const $container = $transaction.find('.subtransaction-container');
+		replaceDisplayContentsAjaxRequest(endpoint, rawData, $container);
+		$transaction.addClass('selected');
+		// Show the expanded transaction summary
+		const $expandedRow = $transaction.find('.expanded');
+		$expandedRow.delay(fadeTime).slideDown(slideTime, function() {
+			$(this).fadeTo(fadeTime, 1);
+		});
 	});
 
 	$iconsLessInfo.on('click', function() {
+		// Get the transaction object
 		const $transaction = $(this).closest('.transaction');
-		$transaction.removeClass('expanded');
+		$transaction.removeClass('selected');
+		// Hide the expanded transaction summary
+		const $expandedRow = $transaction.find('.expanded');
+		$expandedRow.fadeTo(fadeTime, 0, function() {
+			$(this).slideUp(slideTime);
+		})
+		// Show the condensed transaction summary
+		const $condensedRow = $transaction.find('.condensed');
+		$condensedRow.slideDown(slideTime, function() {
+			$(this).fadeTo(fadeTime, 1);
+		});
 	});
 
 })();