+++ CHANGELOG.md
@@ -34,3 +34,10 @@
 - Renamed database files to end with `.sqlite` extension
 - Moved menus on 'Statement Details' and 'Account Details' pages into the sidebar to prevent overlap on collapsed screens
 - Renamed `show_*` route functions to `load_*` to clarify that they are for loading the associated pages (as opposed to just displaying content)
+
+
+## 1.0.5 (in progress)
+
+- Transactions may be split into an arbitrary number of subtransactions, each with a separate subtotal and note
+- Backend properly handles duplicate/ancestor tags; all tags are saved in the database, but only the lowest-level child tag must be entered
+